fix: inherit Entity class

+++ src/Modules/Tours/Explorer.Tours.API/Dtos/QuestionResultDto.cs
@@ -8,8 +8,8 @@ namespace Explorer.Tours.API.Dtos
 {
     public class QuestionResultDto
     {
-        public int QuestionId { get; set; }
-        public int SelectedOptionId { get; set; }
+        public long QuestionId { get; set; }
+        public long SelectedOptionId { get; set; }
         public bool IsCorrect { get; set; }
         public string Feedback { get; set; }
     }
