+++ src/Modules/Tours/Explorer.Tours.API/Dtos/QuizSubmissionDto.cs
@@ -8,7 +8,7 @@ namespace Explorer.Tours.API.Dtos
 {
     public class QuizSubmissionDto
     {
-        public int QuizId { get; set; }
+        public long QuizId { get; set; }
         public long TouristId { get; set; }
         public List<QuizAnswerDto> Answers { get; set; }
     }
